Modified QAOA in noise and varying circuit depth
In this program parameter transfer and selective layer optimization in QAOA is used for different circuit depth. In single circuit depth there is only 1 layer so we are forced to optimize this layer only because no other layer is present. But for circuit depth more than 1, we optimize 2nd layer only and keep all the other layers untouched. There optimization is done after parameters transfer from 6-nodes u3R graph which was fully optimized into p layers of QAOA. So we transferred these 2p optimized parameters into the QAOA for donor graph which is 10-nodes u3R graph as initial parameters and then we selectively optimize them.